Revert relocation and test updates
