+++ src/platform/forms/save-in-progress/ApplicationStatus.jsx
@@ -7,7 +7,7 @@ import {
   formLinks,
   formDescriptions,
   formBenefits,
-} from '../../../applications/personalization/dashboard/helpers';
+} from '../../../applications/personalization/profile360/util/helpers';
 import LoadingIndicator from '@department-of-veterans-affairs/formation/LoadingIndicator';
 import ProgressButton from '@department-of-veterans-affairs/formation/ProgressButton';
 import Modal from '@department-of-veterans-affairs/formation/Modal';
