+++ src/platform/forms/save-in-progress/SaveInProgressIntro.jsx
@@ -9,7 +9,7 @@ import { getNextPagePath } from 'us-forms-system/lib/js/routing';
 import {
   formDescriptions,
   formBenefits,
-} from '../../../applications/personalization/dashboard/helpers';
+} from '../../../applications/personalization/profile360/util/helpers';
 import { toggleLoginModal } from '../../site-wide/user-nav/actions';
 import { fetchInProgressForm, removeInProgressForm } from './actions';
 import FormStartControls from './FormStartControls';
